kernel, vfs, heap, sync, userspace: move fd table to userspace, close heap IRQ deadlock, simplify boot.log dump
Moving the kernel fd table and fstat/dup/dup2/ftruncate into a userspace shim lets the kernel's VFS stay stateless while the libc layer owns offset tracking and sharing semantics. The previous PreemptMutex for HEAP still allowed IRQs to fire while the lock was held, so a timer ISR calling schedule() that missed the per-CPU cache would spin on HEAP with IF=0 forever, and the old allocate_pages helper held HEAP across Vmm::map shootdown_tlb which waits for IPI acks with IRQs disabled creating a shootdown-vs-HEAP inversion on the same CPU. The boot.log dump had deeply nested retry logic with duplicated IOMMU fallback and unispace retry paths, and the PreemptMutex guard was using the pre-lock preemption state rather than tracking whether preemption was actually disabled by try_current_per_cpu.

== Heap IRQ deadlock fix kernel src mm heap rs 1081 kernel src mm heap rs 1285 ==
- Changes HEAP from PreemptMutex to IrqMutex so both preemption and IRQs are disabled while held, closing the self-deadlock where an IRQ handler allocates from the heap misses the per-CPU cache and spins on HEAP held by the interrupted thread on the same CPU.
- Adds heap_grow which reserves VA under HEAP then drops HEAP before the Vmm::map shootdown_tlb wait, doing phys alloc with only preemption disabled and IRQs enabled so shootdown can be acked, then re-acquires HEAP to commit or rollback, serializing concurrent grows via manual preempt_disable without holding the spin lock.
- Alloc drops HEAP before calling heap_grow and re-acquires it after, keeping the lock off during the potentially blocking shootdown path while preserving the chunk reclaim mark_live call under the lock.

== PreemptGuard fix kernel src sync mod rs 22 kernel src sync mod rs 99 ==
- PreemptMutex and PreemptRwLock now track did_disable as a local flag set when try_current_per_cpu succeeds, passing did_disable instead of the pre-lock was_enabled to the guard, so the drop path only re-enables preemption when it was actually disabled by this lock acquisition.
- Fixes the case where try_current_per_cpu returns None during early init or on a CPU without PerCpu allocated, where the guard would previously see was_enabled as true from the pre-lock check and then incorrectly call preempt_enable_and_maybe_resched on drop.
== VFS fd table moved to userspace kernel src filesystems vfs mod rs 853 kernel src filesystems vfs fdtable.rs kernel src filesystems vfs file.rs ==
- Removes kernel-side fstat dup dup2 and ftruncate from the VFS module, deletes the fdtable.rs and file.rs files entirely since the kernel no longer tracks per-task file descriptors.
- The VFS state is now purely path-based: open/read/write/close operate on paths directly and offset tracking is owned by the userspace libc fd layer.

== Userspace fd layer overhaul user libc src fd.rs ==
- Introduces a reference-counted Desc struct for open-file descriptions with path offset readable writable append sync and refcnt fields, plus a separate FdEntry struct that points to a Desc via index so dup shares the description and offset as POSIX requires.
- Adds O_NOFOLLOW via readlink no-follow check, O_DIRECTORY via stat kind check, and O_CLOEXEC tracked per fd entry with F_GETFD F_SETFD properly returning and setting the cloexec bit.
- Splits chunked_write into chunked_write_append using flag 0x1 and chunked_write_positioned using flag base offset shifted left 8, both using the static SCRATCH buffer for the write syscall that consumes its buffer in place.
- Implements fcntl F_DUPFD with the arg hint by scanning for lowest free fd >= arg, F_GETFD returns cloexec bit, F_SETFD sets it, F_GETFL reconstructs flags from desc fields, and F_SETFL sets append and sync.

== Boot log simplification kernel src task load rs 425 ==
- Replaces the deeply nested open-write-close-retry logic with a write_all closure that takes a use_trunc bool and returns bool, using pwrite for TRUNC on the first chunk and positioned writes for subsequent chunks, with a single IOMMU disable and retry path and a single unispace fallback.
- The closure handles the write loop once, the caller tries with TRUNC then retries after IOMMU disable then falls back to unispace, eliminating the duplicated retry blocks.

+++ user/libc/src/stdio.rs
@@ -406,16 +406,16 @@ pub extern "C" fn fseek(f: *mut FILE, offset: c_long, whence: c_int) -> c_int {
         fref.eof = false;
         match whence {
             0 => {
-                fref.offset = offset.max(0) as u64;
+                fref.offset = (offset as i64).max(0) as u64;
                 0
             }
             1 => {
-                fref.offset = (fref.offset as i64 + offset).max(0) as u64;
+                fref.offset = (fref.offset as i64 + offset as i64).max(0) as u64;
                 0
             }
             2 => match file_size(fref) {
                 Some(base) => {
-                    fref.offset = (base as i64 + offset).max(0) as u64;
+                    fref.offset = (base as i64 + offset as i64).max(0) as u64;
                     0
                 }
                 None => {
